navbar finished
I struggled with this part a lot. I used the comparison to see how it's supposed to look like but after hours of trying to figure it out I gave up and just looked at the way that it was written as a reference to the second part to this.